pyproject.toml
[project]
name = "api-playwright-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "allure-pytest>=2.15.3",
    "playwright>=1.58.0",
    "pytest>=9.0.3",
    "pytest-playwright>=0.7.2",
]

# ----------------------------
# PYTEST CONFIG
# ----------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--alluredir=allure-results --clean-alluredir"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
